이벤트 핸들러 순환참조 문제 해결
