+++ service/src/main/java/crypto/trade/eventhandler/LimitSellOrderCreateEventHandler.java
@@ -5,24 +5,22 @@ import crypto.event.EventType;
 import crypto.event.payload.LimitOrderCreateEventPayload;
 import crypto.order.Order;
 import crypto.order.OrderQueryService;
-import crypto.order.OrderService;
 import crypto.time.TimeProvider;
-import crypto.trade.TradeService;
+import crypto.trade.TradeProcessor;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-import static crypto.order.OrderSide.BUY;
 import static crypto.order.OrderSide.SELL;
 
 
 @Component
 @RequiredArgsConstructor
 public class LimitSellOrderCreateEventHandler implements EventHandler<LimitOrderCreateEventPayload> {
-    private final TradeService tradeService;
-    private final OrderService orderService;
+    private final TradeProcessor tradeProcessor;
     private final OrderQueryService orderQueryService;
     private final TimeProvider timeProvider;
 
@@ -31,7 +29,7 @@ public class LimitSellOrderCreateEventHandler implements EventHandler<LimitOrder
         LocalDateTime registeredDateTime = timeProvider.now();
         LimitOrderCreateEventPayload payload = event.getPayload();
 
-        Order sellOrder = orderService.findOrder(payload.getOrderId());
+        Order sellOrder = orderQueryService.findOrder(payload.getOrderId());
         List<Order> buyOrders = orderQueryService.getMatchedLimitSellOrders(sellOrder.getCoin(), SELL, sellOrder.getPrice());
 
         if (buyOrders.isEmpty()) {
@@ -39,7 +37,7 @@ public class LimitSellOrderCreateEventHandler implements EventHandler<LimitOrder
         }
 
         for (Order buyOrder : buyOrders) {
-            tradeService.processMatchLimitOrder(sellOrder, buyOrder, SELL, registeredDateTime);
+            tradeProcessor.processMatchLimitOrder(sellOrder, buyOrder, SELL, registeredDateTime);
 
             if (buyOrder.isFullyFilled()) {
                 sellOrder.markCompleted();
