+++ service/src/main/java/crypto/trade/eventhandler/MarketBuyOrderCreateEventHandler.java
@@ -5,12 +5,12 @@ import crypto.event.EventType;
 import crypto.event.payload.MarketBuyOrderCreateEventPayload;
 import crypto.order.Order;
 import crypto.order.OrderQueryService;
-import crypto.order.OrderService;
 import crypto.time.TimeProvider;
 import crypto.trade.Trade;
 
-import crypto.trade.TradeService;
+import crypto.trade.TradeProcessor;
 import crypto.trade.TradeSettlementService;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -28,8 +28,7 @@ import static crypto.order.OrderSide.SELL;
 @Component
 @RequiredArgsConstructor
 public class MarketBuyOrderCreateEventHandler implements EventHandler<MarketBuyOrderCreateEventPayload> {
-    private final TradeService tradeService;
-    private final OrderService orderService;
+    private final TradeProcessor tradeProcessor;
     private final OrderQueryService orderQueryService;
     private final TradeSettlementService tradeSettlementService;
     private final TimeProvider timeProvider;
@@ -39,7 +38,7 @@ public class MarketBuyOrderCreateEventHandler implements EventHandler<MarketBuyO
         LocalDateTime registeredDateTime = timeProvider.now();
         MarketBuyOrderCreateEventPayload payload = event.getPayload();
 
-        Order buyOrder = orderService.findOrder(payload.getOrderId());
+        Order buyOrder = orderQueryService.findOrder(payload.getOrderId());
         List<Order> sellOrders = orderQueryService.getMatchedMarketBuyOrders(buyOrder.getCoin(), SELL);
 
         BigDecimal remainPrice = buyOrder.getMarKetTotalPrice();
@@ -53,20 +52,20 @@ public class MarketBuyOrderCreateEventHandler implements EventHandler<MarketBuyO
             if (matchedQty.compareTo(BigDecimal.ZERO) <= 0) break;
 
             BigDecimal matchedAmount = sellPrice.multiply(matchedQty);
-            BigDecimal takerFee = tradeService.calculateTradeFee(matchedAmount, TAKER);
-            BigDecimal makerFee = tradeService.calculateTradeFee(matchedAmount, MAKER);
+            BigDecimal takerFee = tradeProcessor.calculateTradeFee(matchedAmount, TAKER);
+            BigDecimal makerFee = tradeProcessor.calculateTradeFee(matchedAmount, MAKER);
             BigDecimal takerTotalUsed = matchedAmount.add(takerFee);
             BigDecimal makerTotalUsed = matchedAmount.add(makerFee);
 
             if (takerTotalUsed.compareTo(remainPrice) > 0) break;
 
-            Trade trade = tradeService.createAndSaveTrade(buyOrder, sellOrder, sellPrice, matchedQty, BUY, takerFee, makerFee, registeredDateTime);
-            tradeService.settleAndMarkOrders(buyOrder, sellOrder, matchedQty, takerTotalUsed, makerTotalUsed, trade, BUY);
+            Trade trade = tradeProcessor.createAndSaveTrade(buyOrder, sellOrder, sellPrice, matchedQty, BUY, takerFee, makerFee, registeredDateTime);
+            tradeProcessor.settleAndMarkOrders(buyOrder, sellOrder, matchedQty, takerTotalUsed, makerTotalUsed, trade, BUY);
             remainPrice = remainPrice.subtract(takerTotalUsed);
         }
 
         if (remainPrice.compareTo(BigDecimal.ZERO) > 0) {
-            BigDecimal totalPrice = remainPrice.add(tradeService.calculateTradeFee(buyOrder.getMarKetTotalPrice(), TAKER));
+            BigDecimal totalPrice = remainPrice.add(tradeProcessor.calculateTradeFee(buyOrder.getMarKetTotalPrice(), TAKER));
             tradeSettlementService.refundUnmatchedLockedBalance(buyOrder, totalPrice);
         }
